// backend/src/common/enums/unit-of-measure.enum.ts
export enum UnitOfMeasure {
  KG = 'KG',           // Kilogram
  GRAM = 'GRAM',       // Gram
  TON = 'TON',         // Metric Ton
  POUND = 'POUND',     // Pound
  
  PALLET = 'PALLET',   // Pallet
  CARTON = 'CARTON',   // Carton/Box
  BAG = 'BAG',         // Bag
  SACK = 'SACK',       // Sack
  
  LITER = 'LITER',     // Liter
  ML = 'ML',           // Milliliter
  GALLON = 'GALLON',   // Gallon
  
  PIECE = 'PIECE',     // Individual pieces
  DOZEN = 'DOZEN',     // Dozen
  
  // Cold storage specific
  CONTAINER = 'CONTAINER', // Shipping container
  TRAY = 'TRAY',          // Storage tray
}